refactoring: graphql testing
* 각 graphql test에 GQLServerTestHelper 및 sequelizeTestHelper 적용
* change file directory graphqlTestClient.js
"spec/graphql/..." -> "spec/testHelper/..."
