+++ backend/spec/graphql/emoji.test.js
@@ -1,30 +1,26 @@
+import assert from "assert";
 import {after, before, describe, it} from "mocha";
-import {GraphQLServer} from "graphql-yoga";
-import GQLClient from "./graphqlTestClient.js";
+import GQLClient from "../testHelper/graphqlTestClient.js";
 import typeDefs from "../../graphQL/typeDefs.js";
 import resolvers from "../../graphQL/resolvers.js";
+import GQLServerTestHelper from "../testHelper/GQLServerTestHelper.js";
+import SequelizeTestHelper from "../testHelper/SequelizeTestHelper.js";
 import config from "../../graphQL/config.js";
-import models from "../../DB/models";
 
 describe("graphql yoga emoji model", () => {
-	let app = null;
+	const gqlServerMock = new GQLServerTestHelper({
+		typeDefs,
+		resolvers,
+		config,
+	});
+	const sequelizeMock = new SequelizeTestHelper();
 
 	before(async () => {
-		const server = new GraphQLServer({
-			typeDefs,
-			resolvers,
-		});
-
-		const serverPromise = server.start(config);
-		const initDBPromise = models.sequelize.sync();
-
-		const res = await Promise.all([serverPromise, initDBPromise]);
-
-		app = res[0];
+		await Promise.all([gqlServerMock.setup(), sequelizeMock.setup()]);
 	});
 
-	after(() => {
-		app.close();
+	after(async () => {
+		await Promise.all([gqlServerMock.teardown(), sequelizeMock.teardown()]);
 	});
 
 	it("able query emoji", async () => {
@@ -39,7 +35,6 @@ describe("graphql yoga emoji model", () => {
 			}
 		}
 		`;
-
 		const variables = {
 			EventId: 2,
 		};
