--- backend/spec/graphql/graphqlTestClient.js
@@ -1,6 +1,0 @@
-import {GraphQLClient} from "graphql-request";
-
-const endpoint = "http://localhost:8000/graphql";
-
-export default new GraphQLClient(endpoint, {headers: {}});
-
